rec-service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
scipy
joblib
numpy
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy>=2.0
psycopg2-binary
python-dotenv